backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
certifi==2026.2.25
click==8.3.1
distro==1.9.0
dnspython==2.8.0
duckdb==1.4.4
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.111.0
fastapi-cli==0.0.24
groq==0.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.2
openpyxl==3.1.3
orjson==3.11.7
packaging==26.0
pandas==2.2.2
patsy==1.0.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2026.1.post1
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
scikit-learn==1.5.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
statsmodels==0.14.2
threadpoolctl==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
